feat: moved the support form modal
